Better 'link' style buttons
To distinguish better between ‘a’ tags and buttons that perform
link-like actions and buttons that perform interaction actions, the
links now consistently change to a lighter colour on hover and have an
underline. Project buttons on the homepage now look like links and are
a white, left-aligned variation of the standard green link.